Refactor views to use TypeScript and script setup syntax
Updated `Home.vue`, `Login.vue`, and `ProblemList.vue` to use TypeScript and the `<script setup>` syntax for better type safety and conciseness. Adjusted imports and improved form handling in `Login.vue`. Removed unused code in `ProblemList.vue`.
