+++ src/main.ts
@@ -1,7 +1,7 @@
 import {createApp} from 'vue'
 import {createPinia} from 'pinia'
 import App from './App.vue'
-import router from './router'
+import router from './router/index'
 import 'highlight.js/styles/github.css'
 import './assets/theme.css'
