+++ src/views/Home.vue
@@ -1,15 +1,12 @@
 <template>
   <div class="home">
     <h1 class="main-title">Welcome to LeetCode Solutions</h1>
-    <p class="text-lg"> This is the home page of our LeetCode solutions website.</p>
+    <p class="text-lg">This is the home page of our LeetCode solutions website.</p>
     <button class="btn-special">Get Started</button>
   </div>
 </template>
 
-<script>
-export default {
-  name: 'Home'
-}
+<script lang="ts" setup>
 </script>
 
 <style scoped>
